fix code splitting
